Abstracting constraint class
